requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy
psycopg2-binary

python-dotenv

PyJWT
passlib==1.7.4
bcrypt==3.2.2
alembic

pytest
httpx
email-validator
python-multipart
